+++ src/components/Follower/CurrentSituation/WeeklyCurrent.tsx
@@ -1,13 +1,8 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-import {
-  IcArrowRightSmallGray,
-  IcLinkWhite,
-  IcStarMiniGray,
-  IcStarMiniYellow,
-  TestWeekboardStatus,
-} from '../../../assets';
-import { CLICKED_DUMMY, DUMMY } from '../../../constants/Follower/currentConst';
+import { IcArrowRightSmallGray, TestWeekboardStatus } from '../../../assets';
+import { DUMMY } from '../../../constants/Follower/currentConst';
+import SolvedQuestionModal from './SolvedQuestionModal';
 
 interface ClickDailyBoardProps {
   nickname: string;
@@ -22,7 +17,6 @@ const WeeklyCurrent = () => {
   });
 
   const { clickedDate, clickedNickname } = clickedBoard;
-  const { questions } = CLICKED_DUMMY;
 
   const handleClickDailyBoard = ({ nickname, date }: ClickDailyBoardProps) => {
     setClickedBoard({
@@ -31,10 +25,6 @@ const WeeklyCurrent = () => {
     });
 
     // 서버 통신 -> 클릭한 닉네임과 날짜를 기반으로 문제 풀이 조회
-  };
-
-  const handleclickLink = (link: string) => {
-    window.open(link);
   };
 
   return (
@@ -75,45 +65,7 @@ const WeeklyCurrent = () => {
                     {count && <TestWeekboardStatus />}
                     <Date>{date}</Date>
 
-                    {isClickedBoard && (
-                      <ModalContainer>
-                        {questions.map((question, idx) => {
-                          const { name, level, platform, link } = question;
-                          const paintedStarArr = Array(level)
-                            .fill(1)
-                            .concat(Array(5 - level).fill(0));
-
-                          return (
-                            <QuestionContainer key={idx}>
-                              <Name>{name}</Name>
-                              <DetailContainer>
-                                <LvStarContainer>
-                                  {paintedStarArr.map((painted, idx) => {
-                                    return (
-                                      <li key={idx}>
-                                        {painted ? (
-                                          <IcStarMiniYellow />
-                                        ) : (
-                                          <IcStarMiniGray />
-                                        )}
-                                      </li>
-                                    );
-                                  })}
-                                </LvStarContainer>
-
-                                <LinkBtn
-                                  type="button"
-                                  onClick={() => handleclickLink(link)}
-                                >
-                                  <IcLinkWhite />
-                                  <Platform>{platform}</Platform>
-                                </LinkBtn>
-                              </DetailContainer>
-                            </QuestionContainer>
-                          );
-                        })}
-                      </ModalContainer>
-                    )}
+                    {isClickedBoard && <SolvedQuestionModal />}
                   </DailyBoard>
                 );
               })}
@@ -211,71 +163,4 @@ const DailyBoard = styled.div<{ $isClicked: boolean }>`
 const Date = styled.p`
   color: ${({ theme }) => theme.colors.white};
   ${({ theme }) => theme.fonts.body_medium_16};
-`;
-
-const ModalContainer = styled.div`
-  display: flex;
-  gap: 1.2rem;
-  align-items: center;
-  flex-direction: column;
-  position: absolute;
-  top: 12.9rem;
-
-  max-height: 50rem;
-  overflow-y: auto;
-
-  z-index: 1;
-
-  padding: 1.2rem;
-
-  border-radius: 0.8rem;
-  background-color: ${({ theme }) => theme.colors.gray700};
-`;
-
-const QuestionContainer = styled.div`
-  display: flex;
-  gap: 1.5rem;
-  flex-direction: column;
-
-  padding: 1.8rem;
-
-  border-radius: 0.8rem;
-  background-color: ${({ theme }) => theme.colors.gray500};
-`;
-
-const Name = styled.p`
-  color: ${({ theme }) => theme.colors.white};
-  ${({ theme }) => theme.fonts.title_bold_20};
-`;
-
-const DetailContainer = styled.div`
-  display: flex;
-  gap: 1.3rem;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const LvStarContainer = styled.ul`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const LinkBtn = styled.button`
-  display: flex;
-  gap: 0.6rem;
-  justify-content: space-between;
-  align-items: center;
-
-  padding: 0.6rem 1.4rem 0.6rem 0.9rem;
-
-  border-radius: 0.6rem;
-  background-color: ${({ theme }) => theme.colors.gray600};
-`;
-
-const Platform = styled.p`
-  color: ${({ theme }) => theme.colors.white};
-  ${({ theme }) => theme.fonts.body_medium_14};
-
-  white-space: nowrap;
 `;
